+++ megalinter/reporters/ConsoleLinterReporter.py
@@ -6,11 +6,8 @@ import logging
 
 import chalk as c
 from megalinter import Reporter, config, utils
-from megalinter.constants import ML_DOC_URL
+from megalinter.constants import ML_DOC_URL_DESCRIPTORS_ROOT
 from megalinter.utils_reporter import log_section_end, log_section_start
-
-mega_linter_version = config.get("BUILD_VERSION", "latest")
-DOCS_URL_DESCRIPTORS_ROOT = f"{ML_DOC_URL}/{mega_linter_version}/descriptors"
 
 
 class ConsoleLinterReporter(Reporter):
@@ -42,7 +39,7 @@ class ConsoleLinterReporter(Reporter):
             )
         else:
             linter_doc_url = (
-                f"{DOCS_URL_DESCRIPTORS_ROOT}/{self.master.descriptor_id.lower()}_"
+                f"{ML_DOC_URL_DESCRIPTORS_ROOT}/{self.master.descriptor_id.lower()}_"
                 f"{self.master.linter_name.lower().replace('-', '_')}"
             )
         # Output linter status