+++ megalinter/reporters/GithubCommentReporter.py
@@ -9,11 +9,8 @@ import re
 
 import github
 from megalinter import Reporter, config
-from megalinter.constants import ML_DOC_URL, ML_REPO_URL
+from megalinter.constants import ML_REPO_URL
 from megalinter.utils_reporter import build_markdown_summary
-
-mega_linter_version = config.get("BUILD_VERSION", "latest")
-DOCS_URL_DESCRIPTORS_ROOT = f"{ML_DOC_URL}/{mega_linter_version}/descriptors"
 
 
 class GithubCommentReporter(Reporter):